+++ PvZ/SnowPeaProjectile.h
@@ -3,7 +3,7 @@
 class SnowPeaProjectile : public PeaProjectile
 {
 public:
-	SnowPeaProjectile(sf::Vector2i vect, int idx_line_, double damage_);
+	SnowPeaProjectile(sf::Vector2f vect, int idx_line_, double damage_);
 	~SnowPeaProjectile();
 	void collisionWithZombies()override;
 };
