+++ PvZ/SnowPeashooter.h
@@ -5,7 +5,7 @@
 class SnowPeashooter : public Peashooter
 {
 public:
-	SnowPeashooter(int idx_line, sf::Vector2i pos);
+	SnowPeashooter(int idx_line, sf::Vector2f pos);
 	~SnowPeashooter();
 
 	void isShooting(double dt)override;
